jonahshop/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.13
django-cors-headers==3.11.0
django-extra-views==0.13.0
django-haystack==3.2.dev0
django-oscar==3.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docutils==0.18.1
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.4.1
factory-boy==3.2.1
Faker==13.3.4
hyperlink==21.0.0
idna==3.3
importlib-resources==5.7.1
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.4.0
kombu==5.2.4
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.46
Pillow==9.1.0
pkg_resources==0.0.0
pluggy==1.0.0
prompt-toolkit==3.0.29
purl==1.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
tblib==1.7.0
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
